[2단계 - 장바구니] 앨버(송상민) 미션 제출합니다

+++ react-shopping-cart/src/components/CartLeftSection/CartLeftSection.jsx
@@ -2,50 +2,38 @@ import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import styled from 'styled-components';
 
+import FlexWrapper from 'components/@shared/FlexWrapper/FlexWrapper';
 import WithSpinner from 'components/@shared/WithSpinner/WithSpinner';
 
-import AllSelectButton from 'components/AllSelectButton/AllSelectButton';
+import AllSelectButton from 'components/AllSelectCheckbox/AllSelectCheckbox';
 import ProductDeleteButton from 'components/ProductDeleteButton/ProductDeleteButton';
-import ShoppingCartItemsContainer from 'components/ShoppingCartItemsContainer/ShoppingCartItemsContainer';
+import ShoppingCartItemGroup from 'components/ShoppingCartItemGroup/ShoppingCartItemGroup';
 
 import {
   deleteCheckedProductsStart,
   fetchCartsStart,
 } from 'redux/carts/carts.action';
 import {
-  selectCartsLoading,
+  selectIsCartsLoading,
   selectCurrentCarts,
 } from 'redux/carts/carts.selector';
 
-import { ColumnFlexWrapper } from 'styles/Wrapper';
-
 import { CURRENT_USER } from 'constants';
 
-const CartLeftSectionHeader = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-`;
-
-const CartItemsContainerHeader = styled.div`
-  font-size: 13px;
-  width: 100%;
-  padding: 20px 0;
-  border-bottom: 2px solid;
-  border-color: ${({ theme }) => theme.colors['gray_01']};
-`;
-
+//재사용X
 function CartLeftSection() {
-  const carts = useSelector(selectCurrentCarts);
   const dispatch = useDispatch();
+  const carts = useSelector(selectCurrentCarts);
+  const isCartsLoading = useSelector(selectIsCartsLoading);
   const myCarts = carts.filter((cart) => cart.user === CURRENT_USER);
-  const cartLoading = useSelector(selectCartsLoading);
   const checkedIdList = myCarts
     .filter((cart) => cart['checked'])
     .map((cart) => cart.id);
 
   const handleDeleteCheckedProducts = () => {
-    dispatch(deleteCheckedProductsStart(checkedIdList));
+    if (window.confirm('선택하신 상품(들)을 삭제하시겠습니까?')) {
+      dispatch(deleteCheckedProductsStart(checkedIdList));
+    }
   };
 
   useEffect(() => {
@@ -53,21 +41,40 @@ function CartLeftSection() {
   }, [dispatch]);
 
   return (
-    <ColumnFlexWrapper width="490px">
-      <CartLeftSectionHeader>
-        <AllSelectButton />
+    <FlexWrapper flexDirection="column" width="490px">
+      <Styled.CartLeftSectionHeader>
+        <AllSelectButton carts={myCarts} />
         <ProductDeleteButton onClick={handleDeleteCheckedProducts}>
           상품삭제
         </ProductDeleteButton>
-      </CartLeftSectionHeader>
-      <div style={{ width: 'inherit' }}>
-        <CartItemsContainerHeader>{`든든배송상품 ${myCarts.length}개`}</CartItemsContainerHeader>
-        <WithSpinner loading={cartLoading}>
-          <ShoppingCartItemsContainer carts={myCarts} />
+      </Styled.CartLeftSectionHeader>
+      <Styled.CartItemsContainer>
+        <Styled.CartItemsContainerHeader>{`든든배송상품 ${myCarts.length}개`}</Styled.CartItemsContainerHeader>
+        <WithSpinner isLoading={isCartsLoading}>
+          <ShoppingCartItemGroup carts={myCarts} />
         </WithSpinner>
-      </div>
-    </ColumnFlexWrapper>
+      </Styled.CartItemsContainer>
+    </FlexWrapper>
   );
 }
+
+const Styled = {
+  CartItemsContainer: styled.div`
+    width: inherit;
+  `,
+
+  CartLeftSectionHeader: styled.div`
+    display: flex;
+    width: 100%;
+    justify-content: space-between;
+  `,
+
+  CartItemsContainerHeader: styled.div`
+    font-size: 13px;
+    width: 100%;
+    padding: 20px 0;
+    border-bottom: 2px solid ${({ theme }) => theme.colors.gray_01};
+  `,
+};
 
 export default CartLeftSection;
